# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Les Casaniers API",
        "description": "Documentation de l'API pour le projet Les Casaniers",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur Principal"
        }
    ],
    "paths": {
...
        "/api/produits": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Liste des produits avec filtres",
                "operationId": "c484df29ca10df7694b945cf5e215340",
                "parameters": [
                    {
                        "name": "categorie_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type_produit",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Créer un produit (Admin)",
                "operationId": "8ea9b3dfcec0aad387a97eecbe585525",
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/produits/{slug}": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Fiche produit détaillée par slug",
                "operationId": "94a07e8af5583a24e6c7f197a1bd3602",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            }
        },
        "/api/produits/{id}": {
            "put": {
                "tags": [
                    "Produits"
                ],
                "summary": "Mettre à jour un produit (Admin)",
                "operationId": "b2d7cf5d440f2ff4ae283c69b79d26b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produits"
                ],
                "summary": "Supprimer un produit (Admin)",
                "operationId": "9d7ef95b330de9729cb8a8f2b393ef88",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/produits/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "Produits"
                ],
                "summary": "Activer/Désactiver un produit (Admin)",
                "operationId": "5cd9e8a5c6dc705e2fce90e904621aae",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}